+++ changelog.md
@@ -8,7 +8,8 @@
 * Movement
   * Added Terrain Slowdown
     * Based on the terrain you're walking on, you'll get slowed down a bit
-  * 
+  * Added Backwards Slowdown
+    * Walking backwards will slowdown the player by 25%
 * Added Passive exhaustion, each second players will get 0.005 exhaustion
 * Tagging has now config options. Also increased slowdown duration (0.25 -> 0.30 seconds for each half heart of damage taken)
 * Data Pack